add neoclip
It looks kinda verbose. And with an overlay it may be kinda more
concise. But in fact we can't apply another overlay before using `pkgs`
in an overlay, since we have an already configerud instance of `pkgs`
passed as an argument to `overlay` function.

So, we just use neoclip package from inputs. Which is straight-forward
and clear.